services/mmaction-recognition/requirements.lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.lock.txt --python-version 3.11 --python-platform linux --no-build-isolation
addict==2.4.0
    # via
    #   mmcv
    #   mmengine
aliyun-python-sdk-core==2.16.0
    # via
    #   aliyun-python-sdk-kms
    #   oss2
aliyun-python-sdk-kms==2.16.5
    # via oss2
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   model-index
    #   opendatalab
    #   openmim
    #   uvicorn
colorama==0.4.6
    # via
    #   opendatalab
    #   openmim
contourpy==1.3.3
    # via matplotlib
crcmod==1.7
    # via oss2
cryptography==46.0.6
    # via aliyun-python-sdk-core
cycler==0.12.1
    # via matplotlib
decord==0.6.0
    # via mmaction2
einops==0.8.2
    # via mmaction2
fastapi==0.115.8
    # via -r requirements.in
filelock==3.14.0
    # via
    #   openxlab
    #   torch
    #   triton
fonttools==4.62.1
    # via matplotlib
fsspec==2026.3.0
    # via torch
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via torch
jmespath==0.10.0
    # via aliyun-python-sdk-core
kiwisolver==1.5.0
    # via matplotlib
markdown==3.10.2
    # via model-index
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via
    #   mmaction2
    #   mmengine
mdurl==0.1.2
    # via markdown-it-py
mmaction2==1.2.0
    # via -r requirements.in
mmcv==2.1.0
    # via -r requirements.in
mmengine==0.10.7
    # via
    #   -r requirements.in
    #   mmcv
model-index==0.1.11
    # via openmim
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.1.1
    # via
    #   -r requirements.in
    #   contourpy
    #   decord
    #   matplotlib
    #   mmaction2
    #   mmcv
    #   mmengine
    #   opencv-contrib-python
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   scipy
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
opencv-contrib-python==4.13.0.92
    # via mmaction2
opencv-python==4.13.0.92
    # via
    #   mmcv
    #   mmengine
opencv-python-headless==4.11.0.86
    # via -r requirements.in
opendatalab==0.0.10
    # via openmim
openmim==0.3.9
    # via -r requirements.in
openxlab==0.1.3
    # via opendatalab
ordered-set==4.1.0
    # via model-index
oss2==2.17.0
    # via openxlab
packaging==24.2
    # via
    #   matplotlib
    #   mmcv
    #   openxlab
pandas==3.0.1
    # via openmim
pillow==12.1.1
    # via
    #   matplotlib
    #   mmaction2
    #   mmcv
    #   torchvision
pip==26.0.1
    # via openmim
platformdirs==4.9.4
    # via yapf
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via
    #   opendatalab
    #   oss2
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
pygments==2.20.0
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.2
    # via uvicorn
pytz==2023.4
    # via openxlab
pyyaml==6.0.3
    # via
    #   mmcv
    #   mmengine
    #   model-index
    #   openxlab
    #   uvicorn
requests==2.28.2
    # via
    #   opendatalab
    #   openmim
    #   openxlab
    #   oss2
rich==13.4.2
    # via
    #   mmengine
    #   opendatalab
    #   openmim
    #   openxlab
scipy==1.17.1
    # via mmaction2
setuptools==60.2.0
    # via openxlab
six==1.17.0
    # via
    #   oss2
    #   python-dateutil
starlette==0.45.3
    # via fastapi
sympy==1.13.1
    # via torch
tabulate==0.10.0
    # via openmim
termcolor==3.3.0
    # via mmengine
torch==2.5.1
    # via
    #   -r requirements.in
    #   mmaction2
    #   torchvision
torchvision==0.20.1
    # via -r requirements.in
tqdm==4.65.2
    # via
    #   opendatalab
    #   openxlab
triton==3.1.0
    # via torch
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==1.26.20
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
yapf==0.43.0
    # via
    #   mmcv
    #   mmengine
